+++ server/routes/aws.js
@@ -1,7 +1,7 @@
-import {upload} from "../models/AWS";
 import {uploadImage} from "../controllers/aws";
 const express = require('express');
+const protect = require('../middleware/auth');
 const router = express.Router();
 
-router.route("/upload",).post(upload.single('image',),uploadImage);
+router.route("/upload",).post(protect,uploadImage);
 module.exports = router;
